refactor adguard code
