+++ near-dns-backend/src/.envrc
@@ -1,0 +1,1 @@
+export ADGUARD_PASSWORD=$(pass adguard/admin)
