+++ near-dns-backend/src/adguard/mod.rs
@@ -1,0 +1,2 @@
+pub mod adguard;
+pub use adguard::reconcile;
